Remove limits on output file size.
The use of self-pointers into a vector<> meant that the vector could not
be relocated, which meant that any growth in the file had to fit within the
inital reservation.

Now, instead of a pointer to a header and a pointer to the contents, those
values are computed dynamically from the underlying vector().